Rename script files, eliminate starter data
- Rename script files in `assets/auth` to avoid naming duplication
- Update module dependencies accordingly
- Remove blank cells array from `store.js` which is no longer needed
while using API connection
