+++ assets/scripts/auth/auth-ui.js
