+++ assets/scripts/index.js
@@ -4,7 +4,7 @@
 // const example = require('./example')
 
 const events = require('./events')
-const authEvents = require('./auth/events')
+const authEvents = require('./auth/auth-events')
 
 // use require without a reference to ensure a file is bundled
 // require('./example')
