+++ assets/scripts/store.js
@@ -2,8 +2,6 @@
 
 const store = {
   currentTurn: 'x',
-  cells: ['', '', '', '', '', '', '', '', ''],
-  over: false,
   xWins: 0,
   oWins: 0
 }
